Create the auth-service  as a independent service and enhanced the login signup UI
