+++ client/src/app/auth/auth.component.html
@@ -85,8 +85,50 @@
         (ngSubmit)="isNext = false"
         novalidate
       >
-        <h3 class="text-2xl font-semibold mb-6 text-center ml-0">Create Account</h3>
-       
+        <div>
+          <div>
+            <h3 class="text-2xl font-semibold mb-6 text-center ml-0">
+              Create Account
+            </h3>
+          </div>
+          <!-- <div class="flex justify-between">
+            <h1 class="w-10 h-10 border-2 rounded-full text-center text-white border-green-400">1</h1>
+            <div class="text-green-400">
+              <p class="text-center">User Details</p>
+              <hr class="border-solid-1 w-36">
+            </div>
+            <h1 class="border w-3 h-3 rounded-full mt-5 bg-green-400"></h1>
+            <div [ngClass]="{ 'text-green-400': !isNext }">
+              <p class="text-center">Extra Information</p>
+              <hr class="border-solid-1 w-36">
+            </div>
+            <h1 class="w-10 h-10 border rounded-full text-center" [ngClass]="{ 'text-white border-green-400': !isNext }">2</h1>
+          </div> -->
+          <div class="flex justify-center mb-6">
+            <div class="flex items-center">
+              <div
+                [ngClass]="isNext ? 'bg-green-500 text-white' : 'bg-gray-300'"
+                class="w-8 h-8 flex items-center justify-center rounded-full"
+              >
+                1
+              </div>
+              <div>
+                <p class="text-center text-green-400">User Details</p>
+                <div class="w-24 h-1 bg-green-500 mx-2"></div>
+              </div>
+              <div>
+                <p class="text-center text-gray-600">Extra Info</p>
+                <div class="w-24 h-1 bg-gray-600 mx-2"></div>
+              </div>
+              <div
+                [ngClass]="!isNext ? 'bg-green-500 text-white' : 'bg-gray-600'"
+                class="w-8 h-8 flex items-center justify-center rounded-full"
+              >
+                2
+              </div>
+            </div>
+          </div>
+        </div>
         <div class="space-y-4">
           <div>
             <label for="signup-username">Username</label>
@@ -171,13 +213,42 @@
         (ngSubmit)="registerUser()"
         novalidate
       >
-      <div class="flex justify-between items-center mb-4">
-        <h3 class="text-2xl font-semibold mb-6 text-center">Create Account</h3>
-         <a
-          class="underline cursor-pointer font-semibold mb-4"
-          (click)="isNext = true"
-          >previous page</a>
+        <div>
+          <div>
+            <div class="flex justify-between items-center mb-4">
+              <h3 class="text-2xl font-semibold mb-6 text-center ml-0">
+                Create Account
+              </h3>
+              <a
+                class="cursor-pointer font-semibold mb-4"
+                (click)="isNext = true"
+                >Back</a
+              >
+            </div>
+            <div class="flex justify-center mb-6">
+              <div class="flex items-center">
+                <div
+                  class="w-8 h-8 flex items-center justify-center rounded-full bg-green-500 text-white"
+                >
+                  1
+                </div>
+                <div>
+                  <p class="text-center text-green-400">User Details</p>
+                  <div class="w-24 h-1 bg-green-500 mx-2"></div>
+                </div>
+                <div>
+                  <p class="text-center text-green-400">Extra Info</p>
+                  <div class="w-24 h-1 bg-green-400 mx-2"></div>
+                </div>
+                <div
+                  class="w-8 h-8 flex items-center justify-center rounded-full bg-green-500 text-white"
+                >
+                  2
+                </div>
+              </div>
+            </div>
           </div>
+        </div>
         <div class="space-y-4">
           <div class="flex gap-4">
             <div class="w-1/2">
@@ -243,18 +314,20 @@
           </div>
 
           <div>
-            <label for="email">Gender</label>
-            <input
-              pInputText
+            <label for="gender">Gender</label>
+            <select
               id="gender"
               name="gender"
               [(ngModel)]="user.gender"
               #gender="ngModel"
               required
-              type="text"
-              class="w-full"
-              placeholder="Gender"
-            />
+              class="w-full p-inputtext"
+            >
+              <option value="" disabled selected>Select Gender</option>
+              <option value="MALE">Male</option>
+              <option value="FEMALE">Female</option>
+              <option value="OTHER">Other</option>
+            </select>
             <div
               *ngIf="gender.invalid && gender.touched"
               class="text-red-500 text-sm"
@@ -307,7 +380,7 @@
             type="submit"
             [disabled]="signupForm.invalid"
             class="w-full mt-4 cursor-pointer"
-            label="Sign Up"   
+            label="Sign Up"
           ></button>
         </div>
       </form>
