+++ server/service-auth/.gitattributes
