+++ server/service-auth/src/main/java/com/springboot/healthcare/AuthServiceApplication.java
@@ -1,0 +1,13 @@
+package com.springboot.healthcare;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AuthServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AuthServiceApplication.class, args);
+	}
+
+}
