+++ server/service-auth/src/main/java/com/springboot/healthcare/SpringSecurityApplication.java
@@ -1,0 +1,13 @@
+package com.springboot.healthcare;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringSecurityApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringSecurityApplication.class, args);
+	}
+
+}
