+++ server/service-auth/src/main/java/com/springboot/healthcare/dto/AuthResponse.java
@@ -1,0 +1,41 @@
+package com.springboot.healthcare.dto;
+
+public class AuthResponse {
+
+    private String token;
+    private String username;
+    private String message;
+
+    // constructor
+    public AuthResponse(String token, String username, String message) {
+        this.token = token;
+        this.username = username;
+        this.message = message;
+    }
+
+    // getters and setters
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+}
