+++ server/service-auth/src/main/java/com/springboot/healthcare/repo/DoctorRepo.java
@@ -1,0 +1,11 @@
+package com.springboot.healthcare.repo;
+
+import com.springboot.healthcare.model.Doctor;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface DoctorRepo extends JpaRepository<Doctor, UUID> {
+}
