+++ server/service-auth/src/main/java/com/springboot/healthcare/repo/PatientRepo.java
@@ -1,0 +1,11 @@
+package com.springboot.healthcare.repo;
+
+import com.springboot.healthcare.model.Patient;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface PatientRepo extends JpaRepository<Patient, UUID> {
+}
