+++ server/service-auth/src/main/java/com/springboot/healthcare/repo/StaffRepo.java
@@ -1,0 +1,11 @@
+package com.springboot.healthcare.repo;
+
+import com.springboot.healthcare.model.Staff;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface StaffRepo extends JpaRepository<Staff, UUID> {
+}
