+++ server/service-auth/src/main/java/com/springboot/healthcare/service/MyUserDetailsService.java
@@ -1,0 +1,27 @@
+package com.springboot.healthcare.service;
+
+import com.springboot.healthcare.model.UserPrincipal;
+import com.springboot.healthcare.model.Users;
+import com.springboot.healthcare.repo.UserRepo;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class MyUserDetailsService implements UserDetailsService {
+
+    @Autowired
+    private UserRepo userRepo;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        Users user = userRepo.findByusername(username);
+        if(user == null) {
+            System.out.println("user not found");
+            throw new UsernameNotFoundException("User not found");
+        }
+        return new UserPrincipal(user);
+    }
+}
