+++ server/service-auth/src/test/java/com/springboot/healthcare/SpringSecurityApplicationTests.java
@@ -1,0 +1,13 @@
+package com.springboot.healthcare;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringSecurityApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
