--- server/src/main/java/com/springboot/springsecurity/SpringSecurityApplication.java
@@ -1,13 +1,0 @@
-package com.springboot.springsecurity;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringSecurityApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringSecurityApplication.class, args);
-	}
-
-}
