--- server/src/main/java/com/springboot/springsecurity/config/JwtFilter.java
@@ -1,68 +1,0 @@
-package com.springboot.springsecurity.config;
-
-import com.springboot.springsecurity.service.JWTService;
-import com.springboot.springsecurity.service.MyUserDetailsService;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-@Component
-public class JwtFilter extends OncePerRequestFilter {
-
-    @Autowired
-    private JWTService jwtService;
-
-    @Autowired
-    ApplicationContext applicationContext;
-
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        String authHeader = request.getHeader("Authorization");
-        String token = null;
-        String username = null;
-
-        if (authHeader != null && authHeader.startsWith("Bearer")) {
-            token = authHeader.substring(7);
-            username = jwtService.extractUserName(token);
-        }
-//        //  Extract JWT from Cookie
-//        if (request.getCookies() != null) {
-//            for (jakarta.servlet.http.Cookie cookie : request.getCookies()) {
-//                if (cookie.getName().equals("jwt")) {
-//                    token = cookie.getValue();
-//                    break;
-//                }
-//            }
-//        }
-//
-//        //  Extract username from token
-//        if (token != null) {
-//            username = jwtService.extractUserName(token);
-//        }
-
-        if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
-
-            UserDetails userDetails = applicationContext.getBean(MyUserDetailsService.class).loadUserByUsername(username);
-            if (jwtService.validateToken(token, userDetails)) {
-                UsernamePasswordAuthenticationToken authToken =
-                        new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-                authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-                SecurityContextHolder.getContext().setAuthentication(authToken);
-            }
-        }
-        filterChain.doFilter(request, response);
-
-    }
-}
