--- server/src/main/java/com/springboot/springsecurity/dto/AuthResponse.java
@@ -1,41 +1,0 @@
-package com.springboot.springsecurity.dto;
-
-public class AuthResponse {
-
-    private String token;
-    private String username;
-    private String message;
-
-    // constructor
-    public AuthResponse(String token, String username, String message) {
-        this.token = token;
-        this.username = username;
-        this.message = message;
-    }
-
-    // getters and setters
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-}
