--- server/src/main/java/com/springboot/springsecurity/model/Users.java
@@ -1,147 +1,0 @@
-package com.springboot.springsecurity.model;
-
-import jakarta.persistence.*;
-
-import java.util.Date;
-import java.util.UUID;
-
-@Entity
-@Table(name = "users")
-public class Users {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "id", updatable = false, nullable = false)
-    private UUID id;
-
-
-    @Column(nullable = false, unique = true)
-    private String username;
-
-    private String first_name;
-
-    private String last_name;
-
-    private Date date_of_birth;
-
-    private String gender;
-
-    private String role;
-
-    @Column(unique = true)
-    private String email;
-
-    private String phone_number;
-
-    @Column(nullable = false)
-    private String password;
-
-    private String status;
-
-    // Getters and Setters
-
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    public Date getDate_of_birth() {
-        return date_of_birth;
-    }
-
-    public void setDate_of_birth(Date date_of_birth) {
-        this.date_of_birth = date_of_birth;
-    }
-
-    public String getGender() {
-        return gender;
-    }
-
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
-
-    public String getRole() {
-        return role;
-    }
-
-    public void setRole(String role) {
-        this.role = role;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getFirst_name() {
-        return first_name;
-    }
-
-    public void setFirst_name(String firstName) {
-        this.first_name = firstName;
-    }
-
-    public String getLast_name() {
-        return last_name;
-    }
-
-    public void setLast_name(String lastName) {
-        this.last_name = lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPhone_number() {
-        return phone_number;
-    }
-
-    public void setPhone_number(String phoneNumber) {
-        this.phone_number = phoneNumber;
-    }
-
-    // toString
-    @Override
-    public String toString() {
-        return "Users{" +
-                ", username='" + username + '\'' +
-                ", first_name='" + first_name + '\'' +
-                ", last_name='" + last_name + '\'' +
-                ", date_of_birth=" + date_of_birth +
-                ", gender='" + gender + '\'' +
-                ", role='" + role + '\'' +
-                ", email='" + email + '\'' +
-                ", phone_number='" + phone_number + '\'' +
-                ", password='" + password + '\'' +
-                '}';
-    }
-}
