--- server/src/main/java/com/springboot/springsecurity/repo/DoctorRepo.java
@@ -1,11 +1,0 @@
-package com.springboot.springsecurity.repo;
-
-import com.springboot.springsecurity.model.Doctor;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface DoctorRepo extends JpaRepository<Doctor, UUID> {
-}
