--- server/src/main/java/com/springboot/springsecurity/repo/PatientRepo.java
@@ -1,11 +1,0 @@
-package com.springboot.springsecurity.repo;
-
-import com.springboot.springsecurity.model.Patient;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface PatientRepo extends JpaRepository<Patient, UUID> {
-}
