--- server/src/main/java/com/springboot/springsecurity/repo/StaffRepo.java
@@ -1,11 +1,0 @@
-package com.springboot.springsecurity.repo;
-
-import com.springboot.springsecurity.model.Staff;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface StaffRepo extends JpaRepository<Staff, UUID> {
-}
