--- server/src/main/java/com/springboot/springsecurity/repo/UserRepo.java
@@ -1,13 +1,0 @@
-package com.springboot.springsecurity.repo;
-
-import com.springboot.springsecurity.model.Users;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface UserRepo extends JpaRepository<Users, UUID> {
-    Users findByusername(String username);
-
-    }
