--- server/src/main/java/com/springboot/springsecurity/service/MyUserDetailsService.java
@@ -1,27 +1,0 @@
-package com.springboot.springsecurity.service;
-
-import com.springboot.springsecurity.model.UserPrincipal;
-import com.springboot.springsecurity.model.Users;
-import com.springboot.springsecurity.repo.UserRepo;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class MyUserDetailsService implements UserDetailsService {
-
-    @Autowired
-    private UserRepo userRepo;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Users user = userRepo.findByusername(username);
-        if(user == null) {
-            System.out.println("user not found");
-            throw new UsernameNotFoundException("User not found");
-        }
-        return new UserPrincipal(user);
-    }
-}
