--- server/src/test/java/com/springboot/springsecurity/SpringSecurityApplicationTests.java
@@ -1,13 +1,0 @@
-package com.springboot.springsecurity;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringSecurityApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
